Remove unnecessary files and modify views test and settings
