--- easy_standardtestcase/templates/easy_standardtestcase/post_detail.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<title>Testing</title>
-</head>
-<body>
-
-
-
-
-
-    <div >
-    	<h1>{{ post.function }}</h1>
-        
-            <div >
-                {{ post.final_standardtestcase }}
-            </div>
-        
-        
-        
-    </div>
-
-
-</body>
-</html>>
