--- easy_standardtestcase/templates/easy_standardtestcase/post_list.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<title>Testing</title>
-</head>
-<body>
-
-
-            {% for post in posts %}
-            <h1><a href="{% url 'post_detail' pk=post.pk %}">{{ post.final_standardtestcase }}</a></h1>
-            {% endfor %}
-
-</body>
-</html>>
